% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_kde_plots.R
\name{plot_constrained_kde_density}
\alias{plot_constrained_kde_density}
\title{Plot the fitted density for the constrained KDE model}
\usage{
plot_constrained_kde_density(model, x_seq)
}
\arguments{
\item{model}{The fitted model object}

\item{x_seq}{The sequence of points to estimate the density}
}
\value{
A ggplot object
}
\description{
Plot the fitted density for the constrained KDE model
}
\examples{
set.seed(42)
n <- 100
x <- c(rnorm(0.5 * n, 0, 1), rnorm(0.5 * n, 1.5, 1/3))
h <- 0.6
model <- fit_constrained_kde(x, h)
x_seq <- seq(min(x) - 1, max(x) + 1, length.out = 1000)
p1 <- plot_constrained_kde_density(model, x_seq)
print(p1)
}
