% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_kde_model.R
\name{predict_constrained_kde}
\alias{predict_constrained_kde}
\title{Predict density and moments using the fitted constrained KDE model}
\usage{
predict_constrained_kde(model, x_seq)
}
\arguments{
\item{model}{The fitted model object}

\item{x_seq}{The sequence of points to estimate the density}
}
\value{
A list containing density estimates and first four moments
}
\description{
Predict density and moments using the fitted constrained KDE model
}
\examples{
set.seed(42)
n <- 100
x <- c(rnorm(0.5*n, 0, 1), rnorm(0.5 * n, 1.5, 1/3))
h <- 0.6
model <- fit_constrained_kde(x, h)
x_seq <- seq(min(x) - 1, max(x) + 1, length.out = 1000)
predictions <- predict_constrained_kde(model, x_seq)
print(predictions$density)
print(predictions$moments)
}
